+++ test/unit/fuzz-tools.test.js
@@ -113,6 +113,7 @@ import * as enviroV111 from '../../lib/enviro-v111.js';
 import * as critcareV112 from '../../lib/critcare-v112.js';
 import * as fluidrespV113 from '../../lib/fluidresp-v113.js';
 import * as pulmV114 from '../../lib/pulm-v114.js';
+import * as pulmnodV115 from '../../lib/pulmnod-v115.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -170,6 +171,7 @@ const MODULES = {
   'critcare-v112.js': critcareV112,
   'fluidresp-v113.js': fluidrespV113,
   'pulm-v114.js': pulmV114,
+  'pulmnod-v115.js': pulmnodV115,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];